+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/misc/Measurement.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.misc;
 
 public class Measurement {
