+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/misc/SCARAController.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.misc;
 
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
