+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/TestingTeleOp.java
@@ -1,21 +1,15 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.opmodes;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
-import org.openftc.revextensions2.ExpansionHubMotor;
+import org.firstinspires.ftc.teamcode.drivecontrol.Robot;
+import org.firstinspires.ftc.teamcode.drivecontrol.Vector2d;
 
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-
-import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.openftc.revextensions2.ExpansionHubMotor;
-
-import org.firstinspires.ftc.teamcode.Angle;
+import org.firstinspires.ftc.teamcode.drivecontrol.Angle;
 
 
-@com.qualcomm.robotcore.eventloop.opmode.TeleOp(name = "TeleOp", group = "TeleOp")
-public class TeleOp extends OpMode {
+@com.qualcomm.robotcore.eventloop.opmode.TeleOp(name = "TestingTeleOp", group = "TeleOp")
+public class TestingTeleOp extends OpMode {
     Robot robot;
     public final double DEADBAND_MAG_NORMAL = 0.1;
     public final double DEADBAND_MAG_SLOW_MODE = 0.03;
@@ -26,6 +20,9 @@ public class TeleOp extends OpMode {
 
     double loopStartTime = 0;
     double loopEndTime = 0;
+
+    double startTime = 0;
+    double mathTime = 0;
 
     //hungry hippo (1 servo, 2 positions)
     //foundation grabber - latch (2 servos, 2 positions)
@@ -39,7 +36,7 @@ public class TeleOp extends OpMode {
     Vector2d joystick1, joystick2;
 
     public void init() {
-        robot = new Robot(this, false, false);
+        robot = new Robot(this, true, false);
     }
 
     public void init_loop() {
@@ -52,17 +49,19 @@ public class TeleOp extends OpMode {
 
     public void start () {
         if (willResetIMU) robot.initIMU();
+        startTime = System.currentTimeMillis();
     }
 
     public void loop() {
         loopStartTime = System.currentTimeMillis();
+        mathTime = (loopStartTime-startTime)/1000;
         telemetry.addData("OS loop time: ", loopEndTime - loopStartTime);
 
         robot.updateBulkData(); //read data once per loop, access it through robot class variable
         robot.driveController.updatePositionTracking(telemetry);
 
-        joystick1 = new Vector2d(gamepad1.left_stick_x, -gamepad1.left_stick_y); //LEFT joystick
-        joystick2 = new Vector2d(gamepad1.right_stick_x+((gamepad1.right_stick_x/Math.abs(gamepad1.right_stick_x))*0.1), -gamepad1.right_stick_y); //RIGHT joystick
+        joystick1 = new Vector2d(Math.cos(mathTime), Math.sin(mathTime)); //LEFT joystick
+        joystick2 = new Vector2d(gamepad1.right_stick_x, -gamepad1.right_stick_y); //RIGHT joystick
         slowModeDrive = false;
 
         telemetry.addData("Robot Heading: ", robot.getRobotHeading());
@@ -73,12 +72,15 @@ public class TeleOp extends OpMode {
         telemetry.addData("Left Orientation: ", robot.driveController.moduleLeft.getCurrentOrientation());
         telemetry.addData("Right Orientation: ", robot.driveController.moduleRight.getCurrentOrientation());
 
+        telemetry.addData("StartTime: ", startTime);
+        telemetry.addData("MathTime: ", mathTime);
+
 
 
         //slow mode/range stuffs
         if (gamepad1.left_trigger > 0.1) {
-           // joystick1 = joystick1.scale(0.3);
-           // joystick2 = joystick2.scale(0.4); //was 0.3
+            // joystick1 = joystick1.scale(0.3);
+            // joystick2 = joystick2.scale(0.4); //was 0.3
             joystick1 = joystick1.scale((1-Math.abs(gamepad1.left_trigger))*.75);
             joystick2 = joystick2.scale(1-Math.abs(gamepad1.left_trigger));
             slowModeDrive = true;
