+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tracking/SimplePathFollow.java
@@ -1,4 +1,6 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tracking;
+
+import org.firstinspires.ftc.teamcode.drivecontrol.Robot;
 
 /*
  * Directly control both motors on both modules to move towards a given robot position and orientation
